Refactor out params and add key to loop item
