+++ resources/js/Components/Code/RunnableCodeBlock.jsx
@@ -1,5 +1,6 @@
 import { useReducer } from "react";
 import { CodeBlock, dracula } from "react-code-blocks";
+import RunnableCodeBlockParameter from "./RunnableCodeBlockParameter";
 
 export default function RunnableCodeBlock({ className = '', title, text, previewCallback, params, children, ...props }) {
     let classes = [
@@ -39,39 +40,12 @@ export default function RunnableCodeBlock({ className = '', title, text, preview
         <div className="border border-gray-400 rounded">
             <div>
                 {params && params.map((param) => {
-                    const optionId = Math.random().toString(36).substring(2, 15);
-                    switch (param.type) {
-                        case 'checkbox':
-                            return <div>
-                                <div>{param.label}</div>
-                                {param.options.map((option, optionIndex) => {
-                                    return <div key={optionIndex} className="inline-block">
-                                        <input
-                                            type="checkbox"
-                                            id={param.id + '-' + option}
-                                            name={param.id}
-                                            checked={options[param.id].includes(option)}
-                                            onChange={(e) => updateOptions({ id: param.id, action: e.target.checked ? 'add' : 'remove', value: option })}
-                                        />
-                                        <label htmlFor={param.id + '-' + option} className='ml-1 mr-4'>{option}</label>
-                                    </div>
-                                })}
-                            </div>;
-                        case 'select':
-                            return <div>
-                                <label className="block" for={param.id + '-' + optionId}>{param.label}</label>
-                                <select
-                                    id={param.id + '-' + optionId}
-                                    name={param.id}
-                                    value={options[param.id]}
-                                    onChange={(e) => updateOptions({ id: param.id, action: 'set', value: e.target.value })}
-                                >
-                                    {param.options.map((option, optionIndex) => {
-                                        return <option key={optionIndex} value={option}>{option}</option>
-                                    })}
-                                </select>
-                            </div>;
-                    }
+                    return <RunnableCodeBlockParameter
+                        key={param.id}
+                        param={param}
+                        options={options}
+                        updateOptions={updateOptions}
+                    />
                 })}
             </div>
             <div className="bg-gray-800 font-mono">
